requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pymongo==4.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
passlib==1.7.4